Tidy up imported CI test a bit
