+++ tests/v2prov/tests/custom/operation_etcdsnapshot_inplace_test.go
@@ -37,11 +37,6 @@ func Test_Operation_SetA_Custom_EtcdSnapshotCreationRestoreInPlace(t *testing.T)
 					ETCD: &rkev1.ETCD{
 						DisableSnapshots: true,
 					},
-					MachineGlobalConfig: rkev1.GenericMap{
-						Data: map[string]any{
-							"debug": true,
-						},
-					},
 				},
 			},
 		},
